Split multiplatform tests into separate class
